requirements.txt
# Core RAG
llama-index-core>=0.10.0
llama-index-vector-stores-qdrant>=0.2.0
llama-index-embeddings-openai>=0.1.0
llama-index-llms-openai>=0.1.0
llama-index-retrievers-bm25>=0.1.0
llama-index-postprocessor-cohere-rerank>=0.1.0

# Vector DB
qdrant-client>=1.7.0

# PDF parsing
pymupdf>=1.23.0
pdfplumber>=0.10.0

# Evaluation
ragas>=0.1.0
datasets>=2.0.0

# UI
gradio>=4.0.0

# Utilities
python-dotenv>=1.0.0
pandas>=2.0.0
tqdm>=4.0.0
tenacity>=8.0.0
